+++ .github/generate_labels.py
@@ -17,81 +17,21 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
-""" Collection of utilities to manage mpm project itself.
+""" Utilities to manage GitHub labels to use in issues and PR management.
 """
 
 from pathlib import Path
 
-from boltons.iterutils import flatten
 from simplejson import dumps as json_dumps
 
-from meta_package_manager.managers import pool
-from meta_package_manager.platform import ALL_OS_LABELS
+from meta_package_manager.labels import LABELS
 
 
-def generate_labels():
-    """ Generate GitHub labels to use in issues and PR management.
+def write_labels():
+    """ Write down labels into JSON file.
     """
     json_file = Path(__file__).parent.joinpath(
         '../.github/labels.json').resolve()
-
-    # Format: label name, color, optional description.
-    LABELS = [
-        ("BitBar plugin", '#fef2c0',
-         "Plugin code, documentation or features related to BitBar"),
-        ("bug", '#d73a4a',
-         "Something isn't working"),
-        ("CI/CD", '#dbca13',
-         "Automation and management of the project itself"),
-        ("documentation", '#006b75',
-         "Update to documentation's content or its generation"),
-        ("duplicate", '#cfd3d7',
-         "This issue or pull request already exists"),
-        ("enhancement", '#84b6eb',
-         "Improvement of an existing feature"),
-        ("feature request", '#fbca04',
-         "Something not existing yet that need to be implemented"),
-        ("good first issue", '#7057ff',
-         "A place for newcomers to start contributing"),
-        ("help wanted", '#008672',
-         "Extra attention is needed"),
-        ("can't reproduce", '#fec1c1',
-         "Root cause unlikely to come from the project"),
-        ("question", '#d876e3',
-         "Further information is requested"),
-        ("wont do/fix", '#eeeeee',
-         "This will not be worked on"),
-    ]
-
-    # Define some colors.
-    PLATFORM_COLOR = '#bfd4f2'
-    MANAGER_COLOR = '#bfdadc'
-
-    # Some managers sharing some roots will be grouped together.
-    MANAGER_GROUPS = {
-        'dpkg-like': {'dpkg', 'apt', 'opkg'},
-        'npm-like': {'npm', 'yarn'},
-    }
-
-    # Create one label per platform.
-    for platform_id in ALL_OS_LABELS:
-        LABELS.append((
-            'platform: {}'.format(platform_id), PLATFORM_COLOR,
-            '{}'.format(platform_id)))
-
-    # Create one label per manager. Add mpm as its own manager.
-    non_grouped_managers = set(
-        pool()) - set(flatten(MANAGER_GROUPS.values())) | {'mpm'}
-    for manager_id in non_grouped_managers:
-        LABELS.append((
-            'manager: {}'.format(manager_id), MANAGER_COLOR,
-            '{}'.format(manager_id)))
-
-    # Add labels for grouped managers.
-    for group_label, manager_ids in MANAGER_GROUPS.items():
-        LABELS.append((
-            'manager: {}'.format(group_label), MANAGER_COLOR,
-            ', '.join(sorted(manager_ids))))
 
     # Debug messages.
     for label_name, _, _ in sorted(LABELS):
@@ -111,4 +51,4 @@ def generate_labels():
 
 
 if __name__ == '__main__':
-    generate_labels()
+    write_labels()